+++ lib/src/lib.rs
@@ -57,7 +57,7 @@ mod world;
 pub mod save;
 
 pub use cells::{Coord, State, ALIVE, DEAD};
-pub use config::{Config, NewState, SearchOrder, SkipLevel, Symmetry, Transform};
+pub use config::{Config, NewState, SearchOrder, Symmetry, Transform};
 pub use error::Error;
 pub use search::Status;
 pub use traits::Search;